Factor out a base class for the wpt metadata compilers
The manifest compilers follow a mostly common design so it seems
reasonable to put the common parts into a base class, and allow
subclasses to override the functionaility where required.

Differential Revision: https://phabricator.services.mozilla.com/D24177

bugzilla-url: https://bugzilla.mozilla.org/show_bug.cgi?id=1536804
gecko-commit: 7f7a2ccd90510ef9d36b0408627fd6ee39f3add5
gecko-integration-branch: central
gecko-reviewers: ato
