+++ tools/wptrunner/wptrunner/manifestexpected.py
@@ -3,7 +3,7 @@ from six.moves.urllib.parse import urljoin
 from collections import deque
 
 from wptmanifest.backends import static
-from wptmanifest.backends.static import ManifestItem
+from wptmanifest.backends.base import ManifestItem
 
 import expected
 
@@ -198,7 +198,7 @@ def fuzzy_prop(node):
 
 
 class ExpectedManifest(ManifestItem):
-    def __init__(self, name, test_path, url_base):
+    def __init__(self, node, test_path, url_base):
         """Object representing all the tests in a particular manifest
 
         :param name: Name of the AST Node associated with this object.
@@ -207,13 +207,14 @@ class ExpectedManifest(ManifestItem):
         :param test_path: Path of the test file associated with this manifest.
         :param url_base: Base url for serving the tests in this manifest
         """
+        name = node.data
         if name is not None:
             raise ValueError("ExpectedManifest should represent the root node")
         if test_path is None:
             raise ValueError("ExpectedManifest requires a test path")
         if url_base is None:
             raise ValueError("ExpectedManifest requires a base url")
-        ManifestItem.__init__(self, name)
+        ManifestItem.__init__(self, node)
         self.child_map = {}
         self.test_path = test_path
         self.url_base = url_base
@@ -339,12 +340,12 @@ class DirectoryManifest(ManifestItem):
 
 
 class TestNode(ManifestItem):
-    def __init__(self, name):
+    def __init__(self, node, **kwargs):
         """Tree node associated with a particular test in a manifest
 
         :param name: name of the test"""
-        assert name is not None
-        ManifestItem.__init__(self, name)
+        assert node.data is not None
+        ManifestItem.__init__(self, node, **kwargs)
         self.updated_expected = []
         self.new_expected = []
         self.subtests = {}
@@ -431,12 +432,6 @@ class TestNode(ManifestItem):
 
 
 class SubtestNode(TestNode):
-    def __init__(self, name):
-        """Tree node associated with a particular subtest in a manifest
-
-        :param name: name of the subtest"""
-        TestNode.__init__(self, name)
-
     @property
     def is_empty(self):
         if self._data:
